refactor: move FAQ restyle into shared FAQsPanel
PR feedback: the previous commit duplicated FAQsPanel inside LandingPage/faq.tsx
(+102 lines, fork of the same accordion logic).

Move the b-loud styling into the shared FAQsPanel component so peanut.me LP,
/lp/card, and MDX blog FAQs all render the same way (MDX FAQ.tsx already
documents that it tracks LP styling). LP wrapper reverts to its original form.

Drops framer-motion animation and the dynamic plus-circle/minus-circle icon —
b-loud has no animation and uses a text glyph.

Net diff vs dev: -37 lines.
